commands: Added option value conversion, and moved option validation logic into Request#convertOptions

+++ commands/option.go
@@ -7,7 +7,7 @@ const (
 	Bool    = reflect.Bool
 	Int     = reflect.Int
 	Uint    = reflect.Uint
-	Float   = reflect.Float32
+	Float   = reflect.Float64
 	String  = reflect.String
 )
